% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{add_snpres}
\alias{add_snpres}
\title{Add a snpRes folder to the prsRepo file}
\usage{
add_snpres(dirpath, ncase = 0, ncontrol = 0, n = 0, ...)
}
\arguments{
\item{dirpath}{filepath for the folder to copy over}

\item{ncase}{n cases}

\item{ncontrol}{n controls}

\item{n}{total n (for continous traits)}

\item{...}{Other arguments}
}
\value{
nothing
}
\description{
Add a snpRes folder to the prsRepo file
}
\examples{
\dontrun{
add_snpres("~/snpres/gwas", common_name = "mdd gwas", pmid=1231313,
ncase=14000, control = 14000, comment = "first mdd GWAS by PGC")}
}
