% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_parres}
\alias{collect_parres}
\title{Collects the parRes output from SbayesR}
\usage{
collect_parres()
}
\value{
a tibble
}
\description{
Collects the parRes output from SbayesR
}
\examples{
\dontrun{
collect_parres()
}
}
