% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_new_repo}
\alias{create_new_repo}
\title{Create the folder structure for a prsRepo, using environmental variables}
\usage{
create_new_repo()
}
\value{
nothing
}
\description{
Create the folder structure for a prsRepo, using environmental variables
}
\examples{
\dontrun{
create_new_repo
}
}
