% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_interface.R
\name{delete_metadata_row}
\alias{delete_metadata_row}
\title{Remove a row from the metadata file}
\usage{
delete_metadata_row(snpres)
}
\arguments{
\item{snpres}{name of the snpres file that you want to delete (folder name in
prsRepo)}
}
\value{
nothing
}
\description{
Remove a row from the metadata file
}
\examples{
\dontrun{
delete_metadata_row("scz2022")
}

}
