% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_interface.R
\name{get_snpres_info}
\alias{get_snpres_info}
\title{Calculates liability scale h2 and expected variance explained by PRS
Checks all sumstats in PRS_REPO, and extracts key data. WHere possible
calculates the expected variance explained in out of sample PRS, and converts
the sbayesR heritability estimate to the liability scale.}
\usage{
get_snpres_info()
}
\value{
a tibble
}
\description{
Calculates liability scale h2 and expected variance explained by PRS
Checks all sumstats in PRS_REPO, and extracts key data. WHere possible
calculates the expected variance explained in out of sample PRS, and converts
the sbayesR heritability estimate to the liability scale.
}
\examples{
\dontrun{
get_snpres_info()
}
}
