% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{get_sumstat_info}
\alias{get_sumstat_info}
\title{Extract information about each sumstat used to get a snpRes file}
\usage{
get_sumstat_info()
}
\value{
a tibble
}
\description{
Extract information about each sumstat used to get a snpRes file
}
\examples{
\dontrun{
get_sumstat_info()
}
}
