% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_interface.R
\name{load_metadata_prsrepo}
\alias{load_metadata_prsrepo}
\title{Loads the info on all snpRes files in PrsRep}
\usage{
load_metadata_prsrepo()
}
\value{
a tibble
}
\description{
Loads the info on all snpRes files in PrsRep
}
\examples{
\dontrun{
load_metadata_prsrepo()
}
}
