% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_metadata_row}
\alias{update_metadata_row}
\title{Update an existing row in the metadata}
\usage{
update_metadata_row(snpres, new_row)
}
\arguments{
\item{snpres}{name of snpres file}

\item{new_row}{the new updated row}
}
\value{
nothing
}
\description{
Update an existing row in the metadata
}
\examples{
\dontrun{

update_metadata_row("scz2022", new_row)
}
}
