% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{add_snpres}
\alias{add_snpres}
\title{Add a snpRes folder to the prsRepo file}
\usage{
add_snpres(
  dirpath,
  common_name = "",
  pmid = "1",
  ncase = 0,
  ncontrol = 0,
  n = 0,
  comment = ""
)
}
\arguments{
\item{dirpath}{filepath for the folder to copy over}

\item{common_name}{The name of the sumstat in common language}

\item{pmid}{pmid of nimh}

\item{ncase}{n cases}

\item{ncontrol}{n controls}

\item{n}{total n (for continous traits)}

\item{comment}{free text comment}
}
\value{
nothing
}
\description{
Add a snpRes folder to the prsRepo file
}
\examples{
\dontrun{
add_snpres("~/snpres/gwas", common_name = "mdd gwas", pmid=1231313,
ncase=14000, control = 14000, comment = "first mdd GWAS by PGC")}
}
