% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayesR_default.R
\name{run_sbayes}
\alias{run_sbayes}
\title{Title}
\usage{
run_sbayes(sumstat, name, impute_n = TRUE, archive = TRUE, sbatch = FALSE)
}
\arguments{
\item{sumstat}{filepath for .ma file to be rescaled}

\item{name}{give the output a name}

\item{impute_n}{Should sbayesR impute sample size?}

\item{archive}{send output to Sys.getenv("PRS_ARCHIVE")}

\item{sbatch}{should the function sbatch the commandline code?}
}
\value{
nothing
}
\description{
Title
}
\examples{
\dontrun{
run_sbayes("here/is_my/gwas_sumstat/sumstat.ma", name = "anorexia_2019")
}
}
